fix warnings on windows

+++ scilab/modules/elementary_functions/src/c/fdlibm/s_ceil.c
@@ -65,7 +65,7 @@ static double huge = 1.0e300;
 		    if(j0==20) i0+=1; 
 		    else {
 			j = i1 + (1<<(52-j0));
-			if(j<i1) i0+=1;	/* got a carry */
+			if((int)j<i1) i0+=1;	/* got a carry */
 			i1 = j;
 		    }
 		}
